Increase number of times "Add Question" tooltip is shown
It's very easy to miss the first time if you click too fast, so it should be shown more than once - but not too much more as then it might get annoying for users who prefer using the mouse. Currently, it will show up for the first two times a new question is added.