Show score in a tooltip when hovering over the quiz progress bar
The quiz score is being calculated based on a partially completed quiz.  Thus, until you're near the end, the score tends to be kind of low - which can be discouraging - especially if it's in plain view all the time. The progress bar gives a better indication of how well you're doing, since it takes into account remaining questions.

As for the actual score, it's probably better to hide this information unless the user specifically goes looking for it.  Some usability improvements could be made towards this end - there should be something to alert the user to the existence of the tooltip (perhaps... another tooltip?), and the progress bar itself might be more clear if it had some sort of label (e.g., "Your Progress:"). I'll leave these for future consideration, but for now, we should hide the low score calculated based on incomplete information.