Deleting all bower-installed components from the repository, and changing the bower install path to /public/vendor
There is now an additional setup step required when building the project, which is to run the command 'bower install'. This will be documented in the README later, along with the rest of the required setup (and perhaps integrated with a build system once I get around to it).

Also, the bower install path was changed from /public/javascripts/vendor to be just /public/vendor.  Not everything installed via bower is necessarily JavaScript (for example, Bootstrap is mostly just CSS), so it doesn't make sense for it to be in a folder called "javascripts".  "vendor" is a bit more generic, and is more appropriate for CSS or font packages.