More descriptive error messages in the quiz service.
Previously used to be along the lines of "Error saving quiz: [Object object]". Now, the error object is extracted to obtain the message and stack trace.

TO DO: It's not a good idea to be showing the end user a stack trace. However, as long as this is still in development, it makes life easier for me.