Edit Quiz: Add a drag handle (button) for reordering questions
Previously, it used to be possible to click and drag on the entire question to reorder it in the list. The ability to reorder questions by dragging was called out by changing the default cursor to a move cursor when hovering over a question. However, since reordering is probably not going to be such a frequently-needed operation, from the UI perspective, this is a bit too distracting. Furthermore, as I'm beginning work on implementing a better rich-text editor for the individual questions, this is going to become problematic.