Creating a prepublish script for installing bower components and ckeditor plugins, and add the confighelper plugin.
Installing bower components used to be done via a postinstall script, but this had a couple issues.

First, it is recommended that these sorts of tasks are done as part of prepublish.  See here:

https://npmjs.org/doc/misc/npm-scripts.html

Second, it didn't work on Windows.

Now, the prepublish script (which is now a javascript file executed by node) checks what platform we're running on and executes the bower install command appropriately.

NOTE: The script currently assumes that bower is installed globally if running on Windows.  This will need to be clarified in the README.  Platforms other than Windows should work fine even if bower is not globally installed (it runs the script in node_modules).