Edit Quiz: Animate transitions when reordering questions by dragging
The approach is based on this post:
http://stackoverflow.com/a/13416536/393005

Here's a demo:
http://jsfiddle.net/dNfsJ/

Some minor tweaks were made to make this more generic (no hard-coded IDs, and no need to explicitly define the shadow list in the HTML), and also fix a couple things (need to erase the "position" property on the item after finishing by calling item.css("position", "") in the stop handler, as otherwise the dragged item will be broken when reordering other items later).

TO DO: This breaks one of the golden rules of AngularJS by introducing DOM manipulation into the controller. Later, I should move this into a separate directive (or extend the angular UI sortable directive), but for now I want to save what I've done since it works.