Reverting commit 5c4444c (attempted fix for ckeditor toolbar glitch)
That commit was attempting to fix the occasional glitch where the ckeditor toolbar for a question would remain visible after deleting the question. However, the fix doesn't do anything, so I'm reverting it.

The issue seems to be easy to reproduce consistently at the moment. Just add a new question - this will cause the ckeditor toolbar for the newly added question to pop up (since the question editor field automatically gets focus). However, observe that clicking outside the editor does not make the toolbar go away.  In fact, even if you were to delete the question, the toolbar would still be visible (though it would jump to the top-left corner of the screen). The only way to make the toolbar disappear is by focusing on another question editor - which pops open a new toolbar, but gets rid of the old one. Now, if you click outside, that new toolbar will also go away.

I'm not sure yet how to go about fixing this. I also noticed that the toolbars get added to the DOM (before the closing </body> tag), but they never get cleaned up - even if you delete the associated questions. This will also need to be addressed.