Animations for adding/removing questions
New questions slide in from the left, expand vertically, and fade in. When removing questions, they slide out to the right, shrink vertically, and fade out.

The approach is a bit hacky (timeouts to avoid setting focus to the question field too soon, and encountering issues with the ckeditor toolbar), but the effect is nice.

ISSUE:

Adding questions via the TAB key is a bit wonky at the moment. Since the focus immediately goes to the question editor field, the ckeditor toolbar initializes too soon, which causes it to be displaced to the left. Later, once the editor has finished loading and the animation has finished playing, we set focus to the field again, and it fixes itself - but it still leaves that initial jitter. We need to somehow prevent the toolbar from showing up right away, or otherwise do something hacky with the focus (e.g., set focus to the "Add Question" button initially, and then set it to the question editor only after the question has finished loading and the animation has finished playing).

Another issue is if you add a question using TAB, the document will scroll in the opposite direction than it's supposed to. Seems to work fine with the "Add Question" button, but not the TAB key.