Fix issues mentioned in previous commit
When adding a question using the TAB key, we initially set focus to the Add Question button, and only after the new question has finished loading do we set it to the question editor for the newly-added question. This fixes the issue with the CKEditor toolbar jittering and the document scrolling in the wrong direction.