Fixing hacky workarounds for the test:e2e grunt task
Pointing the grunt-shell-spawn dependency directly to my fork, which contains fixes for orphaned processes, adds support for the :kill target on Windows, and doesn't convert forward slashes to backslashes. This eliminates the need for the hacky workarounds I was using to kill the orphaned selenium server process after finishing running end-to-end tests.