Bring back the grunt test:cleanup task, and increase wait time after starting selenium server
The test:cleanup task is not explicitly executed by any of the other tasks, but it's good to have it available just in case (easier than having to find the PID by port using netstat and having to execute TaskKill.exe).

The increase in wait time is because I encountered a connection refused error once. It may be wise to increase this even more if I ever start running this on a CI server like Travis so I don't get false positives. Or better yet, make the wait command smarter by scanning selenium's output and looking for a message that indicates that it's ready, and only then proceeding with the tests.