Make the navbar menus be clickable on small (but not mobile-small) screens
Previously, if you resized the browser until it was almost mobile-size (except a bit bigger so that the menu was still visible, instead of collapsing into a toggle button), the menu items were not clickable because the <h2> in the center of the navbar came out on top (the title itself does not necessarily have to be long - the element itself has a constant width). Fixed by adjusting the z-index.