Redo the fix for the offcanvas sidebar in such a way that it doesn't break tests
Figured out why it was stacking instead of appearing side by side: it needed to be positioned absolutely (as in offcanvas.less), but the .make-xx-col mixins were overriding the position: absolute property with position: relative. Since we don't need the .make-xx-col until desktop size or larger, those mixins were put inside a media query.

Other than that, the bottom of the sidebar was being cut off because one of the things included in Bootstrap's offcanvas.css sample was overflow-x: hidden, which specifically meant to hide the scrollbar on mobile. However, we don't want the sidebar to get cut off because that causes selenium tests to break when trying to navigate using the sidebar to a question that's toward the bottom of the list (element not clickable at point). I think the scrollbar is not a big deal, and I'm not sure how to fix the tests otherwise, so I got rid of that part.