Get rid of question validation for now (problematic with autosave)
The issue is as soon as you click Add Question, it triggers autosave, and with the validation in place, the save fails.

I would need to implement some sort of "question draft" functionality first. You should still be able to save questions that don't have all the fields completed, but they shouldn't show up when taking the quiz. In the editor, they show up as a "draft" (i.e., there's some visual indication that the question is not complete, and that it's not going to show up when taking the quiz).

Until this happens, we either can't validate questions, or the way saving works would need to change (get rid of autosave), which I would rather not do.