Edit Quiz: Fix quiz saving (call quiz.save only once)
Previously, it looks like I was calling it once per question, which is likely why this was breaking in production (on dev, saving happened too fast for the error to manifest itself)

As a side effect, I'm taking out the calls to question.validate since there's no validation at the moment anyway. If we did want to do validation, we would need to figure out another way since it's async for each question, and we still only want to call quiz.save once at the very end. A promise-based approach would work well here (using Q.all to validate each question, and then save the quiz in the done() handler).