Edit Quiz: Add the Supplemental Information field
Note: This is implemented for Edit Quiz only at this point - i.e., the field is saved, but it doesn't get displayed when taking the quiz.

The Supplemental Information field is intended to show additional information to the user after they submit an answer. It gets shown regardless of whether or not the submission was correct.