Take Quiz: Put the Next button inside the submission result box
This makes the placement more consistent. It's also a bit more strategic, since if the submission is wrong, the Next button appears next to (or underneath) the correct answer and explanation, which increases the likelihood that the user will actually read it, instead of skipping over it.

Also, this removes having to have multiple instances of the Next button physically present in the DOM, which made writing tests a bit tricky to make sure we were finding the right one to click. (There are still multiple instances of the Submit button, however).

Finally, I'm using the flexbox for layout within the submission result box, which greatly simplifies things.