Add ability to delete topics via the Edit menu in the topics sidebar
Deleting a topic also deletes all subtopics and quizzes, making it a potentially dangerous operation. The confirmation dialog should be improved (using a basic bootbox confirm dialog right now). It should show all the topics and all the quizzes that will be deleted. Also, Undo functionality would be a great addition as well.