Remove drag delay on topic tree
Without the drag delay, there is hardly any distinction between clicking on a topic to select it, and beginning a drag operation - single clicking a topic actually triggers the drag event. We accommodate both possibilities by automatically selecting the topic being dragged. (Note: we don't want to have any drag delay as that would hurt the discoverability of the dragging feature.)